Various changes

+++ include/Heap.h
@@ -1,8 +1,8 @@
 #ifndef __HEAP_H__
 #define __HEAP_H__
 
-typedef double heap_elem;
-typedef unsigned long long heap_len;
+typedef unsigned int heap_elem;
+typedef unsigned int heap_len;
 
 
 void Heapsort(heap_elem*, heap_len);
